+++ tests/core/agents/test_base.py
@@ -26,9 +26,9 @@ from ragas.messages import AIMessage
 from ragas.messages import HumanMessage
 
 from datarobot_genai.core.agents.base import BaseAgent
-from datarobot_genai.core.agents.base import extract_history_messages
 from datarobot_genai.core.agents.base import extract_user_prompt_content
 from datarobot_genai.core.agents.base import make_system_prompt
+from datarobot_genai.core.agents.history import extract_history_messages
 
 
 def _make_run_agent_input_from_dicts(messages: list[dict[str, Any]]) -> RunAgentInput:
